Do not close connections on process exit
This change disables the attempts to auto-close on process exit the
connections, that were opened with `odbc_connect()` and were never
closed with `odbc_close()`.

It appeared that such closure attempt make some drivers (Snowflake) to
complain to stderr and some other (DB2, ClickHouse) to crash the
process.

Testing: new test is added that runs the last and leaves the connection
open.