Add data sources introspection function
This PR adds `odbc_list_data_sources()` table function that can be used
to list all ODBC data sources that are registered in the OS.

It is implemented by calling `SQLDataSources` driver manager function in
a similar way to a drivers listing function added in #81.

Testing: existing `list_env` tests are updated to cover new function
without checking the output.